+++ client/src/components/auth/login-form.tsx
@@ -5,7 +5,7 @@ import { MutateFunction, useMutation } from "react-query";
 
 import { login } from "../../api-service";
 import type { RequestError } from "../../api-service/api-helpers/request-error";
-import { AppToaster } from "../../common/toaster";
+import { AppToaster } from "../../common";
 import { ClearButton, LockButton } from "./form-elements";
 const submitDetails = (
   sendDetails: MutateFunction<
